+++ agent/agent.go
@@ -9,10 +9,8 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"net"
 	"net/http"
 	"os"
-	"runtime"
 	"strings"
 	"sync"
 	"time"
@@ -481,31 +479,10 @@ func (a *Agent) addHandler(msgType string, handlerFunc handler.MessageHandlerFun
 // StartStatsd starts up a StatsD listener on the agent, logs an error for any
 // failures.
 func (a *Agent) StartStatsd() {
-	var runStatsd func() error
 	logger.Info("starting statsd server on address: ", a.statsdServer.MetricsAddr)
 
-	// We need to force a TCP connection for Windows. See
-	// https://github.com/sensu/sensu-go/issues/1402
-	if runtime.GOOS == "windows" {
-		runStatsd = func() error {
-			// TODO: https://github.com/sensu/sensu-go/issues/1498
-			conn, err := net.ListenPacket("tcp", a.statsdServer.MetricsAddr)
-			if err != nil {
-				return err
-			}
-			socketFactory := func() (net.PacketConn, error) {
-				return conn, err
-			}
-			return a.statsdServer.RunWithCustomSocket(a.context, socketFactory)
-		}
-	} else {
-		runStatsd = func() error {
-			return a.statsdServer.Run(a.context)
-		}
-	}
-
 	go func() {
-		if err := runStatsd(); err != nil {
+		if err := a.statsdServer.Run(a.context); err != nil {
 			logger.WithError(err).Errorf("error with statsd server on address: %s, statsd listener will not run", a.statsdServer.MetricsAddr)
 		}
 	}()