+++ CHANGELOG.md
@@ -19,6 +19,7 @@ prior to assuming the existence of said check.
 - Prevent panic when verifying if a metric event is silenced.
 - Add logging around the Sensu event pipeline
 - Marked silenced and hooks fields in event as deprecated
+- StatsD listener on Windows is functional
 
 ### Fixed
 - Metrics with zero-values are now displayed correctly